Save state file more eagerly
We built some infrastructure that informs the user about the currently
playing song based on the state file that MPD writes out. Unfortunately,
this state file is updated only periodically, but not immediately in
response to song changes, causing an undesirable delay.
To fix this issue, this change makes sure to write out said state file
eagerly when the currently playing song changed.